Handle blocks properly in let binds
When a let had a block as the next expression the block was inserted
directly without semi-colons, producing an invalid program.
To fix this we change let bind to always have a block as a next
expression (we emit ";" in rust). This prevent using let bind in
expression positions, but it seems it's not used in the RISC-V model.