Playing a bit with Kani
I tried patching the generated rust code manually to make it executable,
and then wrote the equivalent code manually and tried comparing the two
with Kani.
For the addition with immediate, the code verifies that the manual
implementation is equivalent to the generated one in ~30s. It is
basically a look-up in a vector of register, and addition, and a store,
but uses symbolic values everywhere (immediate, source and destination
registers, and initial registers value).