Add initial scaffolding to transform generated Rust code
The Rust code we generate from Sail is pretty crude and does not compile
because it calls lots of intrinsic functions. Some operations also have
different semantic in Sail and in Rust (for instance, a `+` is Sail
should be a `.wrapping_add` in Rust).

This module will make it possible to transform the raw generated code
into Rust code closer to what we want to have, reducing the need for
manually patching code.