Implement support for ```E_struct``` and ```E_struct_update```
This commit transpiles ```E_struct``` and ```E_struct_update``` by creating two new rust types named RsStruct and RsStructAssign. Currently the translation is not fully complete as we don't have a way to infer the type to generate in the context for the structure. This will come as a next commit.