Reset variable ID counter before each function
This commit makes the variable IDs stable within each function. This is
useful because that way when generating a new model with a small Sail
update only the affected functions are modified, making it easier to
review the changes with a diff.

(cherry picked from commit ee5e18327c2bde6a0351f699851742251facc8b9)