Remove enums in argument position for scattered functions
Scattered functions often use enum literal as differentiating pattern.
In Rust this is not accepted as the pattern is not exhaustive.

This commit fixes the issue by removing all enums in argument position.
For that purpose, it introduces a new function transformation called
`enum_in_args_remover`, which is built from the list of enums.