Add support for the remaining tests
This commit port the remaining tests, at the exception of the decoder
which we will handle last. All the tests now compile to valid Rust code
and pass `cargo test`.

The next step would be to actually add real Rust tests that import the
transpiled Sail models, to verify that the Rust semantic matches the
Sail code.