Start working on v2
This commit does a bunch of refactoring to prepare for an overhaul of
the whole code base. The current version is very much tailored to the
specific subset of the RISC-V model we use in Miralis, on only works on
version 0.17.
We want to aim for something more generic and compatible with the latest
sail version.