Sanitize generic types within type expressions
Generic types were not properly sanitized when used within other types.
For instance, Sail's `option(bits(8 * 'n))` was not properly sanitizes
to `Option<BitVector<{8 * N}>>`.

Note that for expressions involving generics we are still blocked on the
unstable `generic_const_exprs` feature. We might want to use typenum to
as an alternative in the future.