Filter out Mem_write_request
Currently Mem_write_request from the RISC-V model does not compile,
because it relies on const generic exprs (for <{ 8 * N}> where N is a
generic).
There are a few path forward to support Mem_write_request:
- Wait for support in Rust, probably not before a year at least
- Use a workaround such as typenum
- Monomorphise all uses of Mem_write_request

Miralis does not depend on Mem_write_request, so for now we simply
ignore it and move on, but we will have to come back later and implement
a fix.