Add support for configuration
This commit generates a configuration struct by inferring the types of
all Sail configuration expression. The configuration is stored in the
Sail context which is passed as argument to all functions.

Do do so we need to first collect all the configuration items throughout
the functions in the call graph, and then use that data to generate the
rust struct.