Uses u128 bits to avoid integer overflow
Sail has a type system that prevents overflows in some cases by
inferring the return types of binary operations, and choosing a
container that is big enough for the result. In Rust we do not have this
flexibility.
This commit leverages the Sail type inference to guess the correct
container type in the Rust code. For now we limit ourselves to u128,
which is enough for the current use cases.