Infers the type of 'sail_ones' or use a conservative estimate
Sometimes types are not known at compile time in Sail, for instance when
the bit width depends on the platform configuration.
In Rust we need to specify all the types upfront, so we have to resort
to a conservative estimate when we can't know for sure the final value.

We have to be careful to preserve the semantic. In this case we always
return a bitvector full of zeroes, but the vector size can be an
overestimate. This means that in our prelude we have to implement the
built-in functions such that they do not solely rely on the generic
parameter when the size is also passed as a normal parameter (as in
sail_ones for instance).