Add unit tests for the softcore PMP
This commit adds the first tests for the softcore, focuses on the PMP as
it is the biggest chunk we have translated so far.
In addition, it corrects the update of the pmpaddr and pmpcfg setters by
using the Sail implementation (i.e. pmpWriteAddrReg and pmpWriteCfgReg)
to ensure only valid bit patterns can be written.