Rename RISC-V model crate to softcore-rv64
In the future we would like to support multiple architectures (rv32, but
also potentially Arm or CHERI variants). So let's set-up a naming
convention right now.
As to why softcore, I like the name. In the end we are packaging a
software core. I know about soft cores on FPGA, but I don't think it is
that big of a deal as the distinction can easily be made from the
context.