compiler: generate doc comments
The compiler now emits doc comments when generating functions, enums,
and struct. The comments contains the location in the Sail sources,
which makes it easy to localize the code and debug potential compiler
issues.

Some of the names generated by the compiler have also been updated, such
as `SailContext` which is now simply `Core` to remove mentions of Sail
from the code directly exposed to the library user.