compiler: remove `Copy` derive from generated core and config
We use copy semantic for all Sail values, as they are mostly small bit
vectors and that makes borrow checking much simpler, however we used to
derive Copy for all structs, including the giant Core structure holding
all the registers of the model.
This commit adds a `derive` fields to structs in the compiler to allow
selecting the traits to derive on a per-struct basis. Using this, we no
longer derive Copy for the Core and Config structs.