compiler: make hoisting more precise
The compiler used to be very eager to hoist expressions, which makes it
harder to read the Rust code and relate it to the Sail source. This
commit makes the hoisting heuristic more precise, in particular it does
not hoist all arguments when only a few needs to be hoisted, like the
previous version did.