compiler: emit top level let as constants
Previously we were only emitting Sail numeral types as Rust constants,
with this commit we additionally create constants for top level lets.
When a type and a let binding have the same name, we only emit the let.

Note that this might cause semantic bugs if the numeral type does not
have the same value as the let binding. We should definitely add a check
somehow that both are equal (maybe emitting new value with a constant
assert?), but for now we simply hope that the Sail models we consider
are reasonable and do not have such foot guns.

Also important: the rv64 model no longer compile with this commit. This
is because some of the newly emitted constants use features that are not
yet supported by our compiler in const contexts.