compiler: mark trivial functions as const
This commit adds a very simple heuristic that marks trivial functions as
const. A function is trivially const if it contains a single literal, or
a single call to a const function from the prelude. In the future we can
refine the heuristic to make it smarter, but for now that is good enough
to fix the compilation errors.