compiler: more aggressive constant propagation
We were getting compilation errors because some of the arithmetic in the
RISC-V model overflows 64 bits. This is especially the case for
constants such as the maximum usize computed as `2 ^ 64 - 1`: although
the final value fits in a usize, the computation overflows when
computing the exponential.

With this commit the compiler will perform constant propagation and
unfolding much more aggressively, but does so using Big_int internally.
This means that the compiler can perform operations that would overflow
a usize as long as the final result fits inside.

In particular, one of the changes is that the compiler will
cross-optimize the constants iteratively, inlining constants into
others once they reach an irreducible literal.