compiler: add support for static config
Some config values needs to be know statically, and can not be read form
the context. This is the case for types who depends on a config value
(like the maximum vector size).

To handle those, this commit adds support for a set of static config
values. For now those are hardcoded in the compiler, but we could image
keeping one config file per Sail model in the future.

Ideally, we would also like to use generics rather than hard-coding the
values in the Rust model, but that would require some support for const
generic. We will see what are the options when some fragments of const
generics get stabilized in the future.