compiler: add support for remaining features required for the decoder
The main change is the handling of `unsigned`, which now returns a i128
integer. Because bitvectors are at most 64 bits interpreted as u64, the
conversion returns a positive integer as expected.

This fixes the remaining integer typing issues, and seems like a better
solution than tentatively tracking types and doing manual conversions on
smaller integer types (u64 to i64).