rv64: add decoder to the RISC-V model
This commit adds support for the decoder in the RISC-V model. This
drastically increases the size of the model, and seems to have an impact
on the time it takes Rust Analyzers to process the raw model. We will
see how this scales when including more of the model.

This commit is a bit of a milestone for softcore-rs, as we now support a
non-negligible chunk of the full RISC-V architecture, yeay!