all: add support for reading/writing general purpose registers
This commit adds support for reading and writing general purpose
registers in the rv64 model. This required adding the rX and wX
functions to the call set, which in turns caused more code generation in
the tests and the additional support for `sail_zero_extend` in our
bitvector transformations.

The rv64 module now expose a simpler way to read/write general purpose
registers through the set/get function.