compiler: link parameters and generic arguments to help type inference
In Sail some generics (i.e. type parameters in Sail parlance) are
deduced from the concrete value of an argument. In Rust that is not
possible, so we have both a generic parameter and an argument.

However, the Rust generic and argument have to be equal. This commit set
the generic to a compile-time known value equal to the argument when
possible. This helps type inference and will enable the Rust type system
to catch some of our compiler bugs (hopefully not too many!).