feat(compiler): support scattered functions with more than one argument
Scattered functions with more than one argument were mis-compiled
because of a confusion between multiple arguments and tuple type due to
the mismatch between the Sail and Rust representation.
This commit fixes the issue by adding a transformation that detects such
miss-compilation and correct the match statement by re-constructing the
appropriate tuple.