feat(compiler): add support for inligning constant functions
All functions that return a fixed literal value can now be inlined by
the optimizer. This allow further optimizing expressions with function
call.
Note that this drops the arguments, and thus assume that they are indeed
pure functions. Ideally we would first check the purity of all arguments
before performing the substitution.