feat(compiler): add support for let binding constant propagation
We already had support for global constant propagation (i.e. top-level
rust const), this commit adds support for propagating constants binded
to a name.

For instance, the following code:

```
let a = 5
if a = 6 { ... }
```

is now optimized to:

```
if 5 = 6 { ... }
```

In turn, this enables more optimizations to kick-in, by removing
indirections through let bindings.