feat(rv64)!: `get_csr` and `read_csr` return options instead of panicking
Previously `get_csr` and `read_csr` were panicking when the CSR was not
known. This turned out to be annoying when used with verification tools
as it would simply cause verification to fail on any unknown CSR. The
new version returns an option instead, offering more control over when
to panic.