feat(armv9a): add support for aarch64 'add_sub_carry'
This commit extends the compiler to support a first Arm instruction.
The required changes focuses on extending the set of external_func for
armv9a, as well as increasing the compiler support of Sail's AST
(notably the addition of Sail's `var` (transpiled to Rust's `let mut`)
and type variables in let bindings.