feat(compiler): remove unused generics in Rust code
Some of Sail's generics get erased when translating to Rust. That is
fine, because Sail already did the type-checking and we know that the
program is correct. However, keeping those generics will confuse the
Rust compiler because it does not know how to infer them.
This commit checks the set of used generic parameters for each
functions, and prune all unused generics.