feat(compiler): add support for "unsupported functions"
Unsupported functions are ignored during the call graph computation, and
all calls to unsupported functions are replaced by `panic!`s by a later
optimisation pass.
This is an important step to support the `execute` function of the
RISC-V model, as it allows to remove unsupported features and
progressively add them back as our Sail support improves.