fix(compiler): use int type instead of nat for big multiplications
We switched to using int128 as the default integer representation, and
cast into u128 only when required. This commit switches the type for the
multiplication that might overflow 64 bits.

We might need to be careful we multiplying two 64 bits numbers to not
loose the sign bit. This is a known limitation of the current
implementation, but so far no Sail model we explored required as many
bits as to cause an overflow.